file for protocolSetting
